Refactored prune_saltdb.py script. Major speed improvement.

+++ scripts/prune_saltdb.py
@@ -5,7 +5,7 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql+psycopg2://salt:salt@127.0.0.1:5432/salt'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql+psycopg2://salt-admin:salt-passwd@127.0.0.1:5432/salt'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 meta = db.metadata
@@ -21,19 +21,14 @@ class Returns(db.Model):
 keepdays = 10
 
 def clean(keepdays):
-    old_returns = Returns.query.filter(Returns.alter_time < datetime.datetime.utcnow() - datetime.timedelta(days=keepdays))
-    try:
-        o = old_returns[0]
+    o = Returns.query.filter(Returns.alter_time < datetime.datetime.utcnow() - datetime.timedelta(days=keepdays)).first()
+    if o is not None:
         tzinfo = o.alter_time.tzinfo
         old_returns = Returns.query.filter(Returns.alter_time < datetime.datetime.now(tzinfo) - datetime.timedelta(days=keepdays))
-        jids = set([Jids.query.filter(Jids.jid == o.jid) for o in old_returns])
-        print("Pruning {0} old jobs and {1} old returns".format(len(jids), old_returns.count()))
-        for job in jids:
-            job.delete()
+        old_jobs = Jids.query.filter(Jids.jid <=o.jid)
+        print("Pruning {0} old jobs and {1} old returns".format(old_jobs.count(), old_returns.count()))
         old_returns.delete()
         db.session.commit()
-    except IndexError:
-        pass
 
 if __name__=='__main__':
     clean(keepdays)
